Refractored Learner Class
There is a Base Learner class called Learner, which
should be extended with implementations of _calculate_losses()
and _train(). The previous implementation is in a new class
called DefaultLearner. Which Learner is used is defined
in the learner_factory.py file, where the
register_learner() function can be called, similar
to a custom model.